+++ packages/create-theme/public/template/docs/.vitepress/blog-theme.ts
@@ -24,7 +24,7 @@ const blogTheme = getThemeConfig({
   // 默认开启pagefind离线的全文搜索支持（如使用其它的可以设置为false）
   // search: false,
 
-  // 默认开启 markdown 图表支持（会增加一定的构建耗时）
+  // 默认关闭 markdown 图表支持（开启会增加一定的构建耗时）
   // mermaid: false
 
   // 页脚