+++ packages/shared/src/index.ts
@@ -1,3 +1,5 @@
 export * from './fs'
 export * from './date'
 export * from './vitepress'
+export * from './utils'
+export * from './getGitTimestamp'